Inventory
Tab to open, not a toggle, you have to hold down tab, makes use of panels with images and button behaviors, in the future we will be able to change which item may go in each slot, with only 5 items we would be able to make each slot dedicated for each item. using perhaps some sorta while clause or "while in inventory" or "while visible" to trigger the effects of an item like the acid proof boots, or make the button an on use item for something like the medkit. the health system should be done next